add reset method and build package
